Update README.md
This update gives a complete readout of the goals and objectives, coding, data wrangling, mathematical and statistical testing/analysis, and visualization results. All essential files containing code is in two areas: Python for random forest and decision tree analysis and R for the more challenging mathematical and statistical testing/analysis along with experimental 3D plotting in RStudio. Tableau was applied for a more in depth multivariable approach to visualization of the data for determining patterns and producing potential models.